fix: remaining codereview refactors
