+++ ui/src/App.js
@@ -2,11 +2,11 @@ import './App.css';
 import Authorize from './pages/Authorize';
 import CreateAccount from './pages/AccountCreation/CreateAccount';
 import Loading from './pages/Loading';
-import SignIn from './pages/SignIn';
+import SignIn from './pages/SignIn/SignIn';
 import Reset from "./pages/Reset";
 import CreateOrRecover from './pages/CreateOrRecover';
 import Recover from './pages/Recovery/Recover';
-import ViewSecretPhrase from './pages/ViewSecretPhrase';
+import ViewSecretPhrase from './pages/ViewPhrase/ViewSecretPhrase';
 import { Container } from 'semantic-ui-react';
 import { appWindow } from '@tauri-apps/api/window';
 import { invoke } from '@tauri-apps/api/tauri';
